+++ user-story/src/components/organisms/NavBar/index.tsx
@@ -1,16 +1,9 @@
 import Typography from '@/components/atoms/Typography'
 import IconTypography from '@/components/molecules/IconTypography'
-import {
-  DASHBOARD,
-  HOME,
-  INFO,
-  RECENT,
-  SETTING,
-  SIDEBAR_DATA,
-} from '@/strings/string'
+import { SIDEBAR_DATA } from '@/strings/string'
+import { handleClick } from '@/utils/function'
 import { Card, Stack, styled } from '@mui/material'
 import { useState } from 'react'
-import { useNavigate } from 'react-router-dom'
 
 const MainContainer = styled(Stack)({
   display: 'flex',
@@ -27,30 +20,7 @@ const Content = styled(Stack)({
   marginLeft: '30%',
 })
 const NavBar = () => {
-  const navigate = useNavigate()
   const [active, setActive] = useState<string>('Home')
-  const handleClick = (label: string) => {
-    setActive(label)
-    switch (label) {
-      case HOME:
-        navigate('/')
-        break
-      case DASHBOARD:
-        navigate('/dashboard')
-        break
-      case INFO:
-        navigate('/info')
-        break
-      case RECENT:
-        navigate('/recent')
-        break
-      case SETTING:
-        navigate('/setting')
-        break
-      default:
-        break
-    }
-  }
   return (
     <MainContainer>
       <NavContainer>
@@ -65,9 +35,6 @@ const NavBar = () => {
           />
         ))}
       </NavContainer>
-      <Content>
-        <Typography variant="h2">{active}</Typography>
-      </Content>
     </MainContainer>
   )
 }
