+++ user-story/src/strings/string.ts
@@ -19,3 +19,8 @@ export const DASHBOARD = 'Dashboard'
 export const INFO = 'Info'
 export const RECENT = 'Recent'
 export const SETTING = 'Setting'
+export const home = '/'
+export const dashboard = '/dashboard'
+export const info = '/info'
+export const recent = '/recent'
+export const setting = '/setting'
